% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_stat_summary.R
\name{calculate_stat_summary}
\alias{calculate_stat_summary}
\title{Create summary information about the covid in each province of Canada}
\usage{
calculate_stat_summary(df, column)
}
\arguments{
\item{df}{data.frame containing covid data to summarize}

\item{column}{the column name to summarize}
}
\value{
data.frame containing summary information.
}
\description{
Create summary information about the covid in each province of Canada
}
\examples{
df <- get_covid_data()
calculate_stat_summary(df,'cases')
}
