% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_covid_data.R
\name{get_covid_data}
\alias{get_covid_data}
\title{Acquires Canada Covid data of specified type
and for an optionally provided date and province}
\usage{
get_covid_data(data_type = "cases", loc = "", date = "")
}
\arguments{
\item{data_type}{string containing type of metric requested}

\item{loc}{string indicating province of interest}

\item{date}{string indicating date of interest}
}
\value{
dataframe of Covid data
}
\description{
Acquires Canada Covid data of specified type
and for an optionally provided date and province
}
\examples{
get_covid_data('cases', 'BC', '13-01-2021')
}
