% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Create a plot of the number of covid cases over time.}
\usage{
plot_timeseries(covid_df, metric, start = "", end = "")
}
\arguments{
\item{covid_df}{a dataframe containing covid data to plot}

\item{metric}{the name of column from covid data to plot along time}

\item{start}{a string indicating the start date in pattern YYYY-MM-DD}

\item{end}{a string indicating the end date in pattern YYYY-MM-DD}
}
\value{
plot object
}
\description{
Create a plot of the number of covid cases over time.
}
\examples{
df <- get_covid_data()
plot_timeseries(df,cases)
}
